Display syntax error also through LibraryLoadingError
