+++ pipelex/builder/concept/concept_spec.py
@@ -15,9 +15,7 @@ from pipelex.core.concepts.concept_blueprint import (
 from pipelex.core.concepts.concept_native import NativeConceptCode
 from pipelex.core.concepts.exceptions import ConceptCodeError, ConceptStringOrConceptCodeError
 from pipelex.core.domains.domain_blueprint import DomainBlueprint
-from pipelex.core.memory.working_memory import WorkingMemory
 from pipelex.core.stuffs.structured_content import StructuredContent
-from pipelex.system.registries.func_registry import pipe_func
 from pipelex.tools.misc.string_utils import is_pascal_case, normalize_to_ascii, snake_to_pascal_case
 from pipelex.types import Self, StrEnum
 
@@ -260,27 +258,3 @@ class ConceptSpec(StructuredContent):
                 converted_structure[field_name] = field_spec.to_blueprint()
 
         return ConceptBlueprint(description=self.description, structure=converted_structure, refines=self.refines)
-
-
-@pipe_func()
-async def create_concept_spec(working_memory: WorkingMemory) -> ConceptSpec:
-    concept_spec_draft = working_memory.get_stuff_as(name="concept_spec_draft", content_type=ConceptSpecDraft)
-    concept_spec_structures_stuff = working_memory.get_stuff_as_list(name="concept_spec_structures", item_type=ConceptStructureSpec)
-
-    structure_dict: dict[str, ConceptStructureSpec] = {}
-    for structure_item in concept_spec_structures_stuff.items:
-        structure_spec = ConceptStructureSpec(
-            the_field_name=structure_item.the_field_name,
-            description=structure_item.description,
-            type=structure_item.type,
-            required=structure_item.required,
-            default_value=structure_item.default_value,
-        )
-        structure_dict[structure_item.the_field_name] = structure_spec
-
-    return ConceptSpec(
-        the_concept_code=concept_spec_draft.the_concept_code,
-        description=concept_spec_draft.description,
-        structure=structure_dict,
-        refines=concept_spec_draft.refines,
-    )
