+++ pipelex/builder/pipe/pipe_parallel_spec.py
@@ -5,8 +5,8 @@ from typing_extensions import override
 
 from pipelex.builder.concept.concept_spec import ConceptSpec
 from pipelex.builder.pipe.pipe_spec import PipeSpec
+from pipelex.builder.pipe.pipe_spec_exceptions import PipeParallelSpecError
 from pipelex.builder.pipe.sub_pipe_spec import SubPipeSpec
-from pipelex.exceptions import PipeDefinitionError
 from pipelex.pipe_controllers.parallel.pipe_parallel_blueprint import PipeParallelBlueprint
 from pipelex.types import Self
 
@@ -46,7 +46,7 @@ class PipeParallelSpec(PipeSpec):
                 "PipeParallel requires either add_each_output to be True or combined_output to be set, "
                 "or both, otherwise the pipe won't output anything"
             )
-            raise PipeDefinitionError(
+            raise PipeParallelSpecError(
                 message=msg,
                 pipe_code=self.pipe_code,
                 description=self.description,
