+++ pipelex/cli/commands/validate_cmd.py
@@ -1,23 +1,24 @@
 from __future__ import annotations
 
 import asyncio
-from typing import Annotated
+from typing import TYPE_CHECKING, Annotated
 
 import click
 import typer
 from rich.console import Console
-from rich.panel import Panel
 from rich.syntax import Syntax
 
 from pipelex import log, pretty_print
 from pipelex.builder.builder_errors import PipelexBundleError
 from pipelex.builder.builder_validation import validate_dry_run_bundle_blueprint
 from pipelex.core.interpreter import PipelexInterpreter
-from pipelex.exceptions import PipeInputError
+from pipelex.exceptions import LibraryLoadingError, PipeInputError
 from pipelex.hub import get_pipes, get_required_pipe
 from pipelex.pipe_run.dry_run import dry_run_pipe, dry_run_pipes
 from pipelex.pipelex import Pipelex
 
+if TYPE_CHECKING:
+    from pipelex.core.validation_errors import ValidationErrorDetailsProtocol
 console = Console()
 
 
@@ -103,7 +104,12 @@ def validate_cmd(
 
     async def validate_pipe(pipe_code: str | None = None, bundle_path: str | None = None):
         # Initialize Pipelex
-        pipelex_instance = Pipelex.make()
+        try:
+            pipelex_instance = Pipelex.make()
+        except LibraryLoadingError as library_loading_error:
+            typer.secho(f"Failed to validate: {library_loading_error}", fg=typer.colors.RED, err=True)
+            present_validation_error(details_provider=library_loading_error)
+            raise typer.Exit(1) from library_loading_error
 
         if bundle_path:
             # When validating a bundle, load_pipe_from_bundle validates ALL pipes in the bundle
@@ -119,37 +125,38 @@ def validate_cmd(
                 raise typer.Exit(1) from exc
             except PipelexBundleError as bundle_error:
                 typer.secho(f"\n❌ Failed to validate bundle '{bundle_path}':", fg=typer.colors.RED, err=True)
-                if concept_definition_errors := bundle_error.concept_definition_errors:
-                    for concept_definition_error in concept_definition_errors:
-                        if syntax_error_data := concept_definition_error.structure_class_syntax_error_data:
-                            message = concept_definition_error.message
-                            code = concept_definition_error.structure_class_python_code or ""
-                            highlight_lines: set[int] | None = None
-                            if syntax_error_data.lineno:
-                                highlight_lines = {syntax_error_data.lineno}
-                            syntax = Syntax(
-                                code=code,
-                                lexer="python",
-                                line_numbers=True,
-                                word_wrap=False,
-                                # theme="monokai",  # pick any theme rich knows; omit to use default
-                                theme="ansi_dark",  # pick any theme rich knows; omit to use default
-                                line_range=None,
-                                highlight_lines=highlight_lines,
-                            )
-                            pretty_range = ""
-                            if syntax_error_data.lineno and syntax_error_data.end_lineno:
-                                pretty_range = f"lines {syntax_error_data.lineno} to {syntax_error_data.end_lineno}"
-                            elif syntax_error_data.lineno:
-                                pretty_range = f"line {syntax_error_data.lineno}"
-                            if syntax_error_data.offset and syntax_error_data.end_offset:
-                                pretty_range += f", column {syntax_error_data.offset} to {syntax_error_data.end_offset}"
-                            elif syntax_error_data.offset:
-                                pretty_range += f", column {syntax_error_data.offset}"
-                            console.print(message)
-                            if pretty_range:
-                                pretty_print(f"Generated code error at {pretty_range}")
-                            console.print(syntax)
+                present_validation_error(details_provider=bundle_error)
+                # if concept_definition_errors := bundle_error.get_concept_definition_errors():
+                #     for concept_definition_error in concept_definition_errors:
+                #         if syntax_error_data := concept_definition_error.structure_class_syntax_error_data:
+                #             message = concept_definition_error.message
+                #             code = concept_definition_error.structure_class_python_code or ""
+                #             highlight_lines: set[int] | None = None
+                #             if syntax_error_data.lineno:
+                #                 highlight_lines = {syntax_error_data.lineno}
+                #             syntax = Syntax(
+                #                 code=code,
+                #                 lexer="python",
+                #                 line_numbers=True,
+                #                 word_wrap=False,
+                #                 # theme="monokai",  # pick any theme rich knows; omit to use default
+                #                 theme="ansi_dark",  # pick any theme rich knows; omit to use default
+                #                 line_range=None,
+                #                 highlight_lines=highlight_lines,
+                #             )
+                #             pretty_range = ""
+                #             if syntax_error_data.lineno and syntax_error_data.end_lineno:
+                #                 pretty_range = f"lines {syntax_error_data.lineno} to {syntax_error_data.end_lineno}"
+                #             elif syntax_error_data.lineno:
+                #                 pretty_range = f"line {syntax_error_data.lineno}"
+                #             if syntax_error_data.offset and syntax_error_data.end_offset:
+                #                 pretty_range += f", column {syntax_error_data.offset} to {syntax_error_data.end_offset}"
+                #             elif syntax_error_data.offset:
+                #                 pretty_range += f", column {syntax_error_data.offset}"
+                #             console.print(message)
+                #             if pretty_range:
+                #                 pretty_print(f"Generated code error at {pretty_range}")
+                #             console.print(syntax)
                 raise typer.Exit(1) from bundle_error
             except PipeInputError as exc:
                 typer.secho(f"\n❌ Failed to validate bundle '{bundle_path}': {exc}", fg=typer.colors.RED, err=True)
@@ -168,3 +175,41 @@ def validate_cmd(
             raise typer.Exit(1)
 
     asyncio.run(validate_pipe(pipe_code=pipe_code, bundle_path=bundle_path))
+
+
+def present_validation_error(details_provider: ValidationErrorDetailsProtocol):
+    concept_definition_errors = details_provider.get_concept_definition_errors()
+    if not concept_definition_errors:
+        return
+    for concept_definition_error in concept_definition_errors:
+        syntax_error_data = concept_definition_error.structure_class_syntax_error_data
+        if not syntax_error_data:
+            continue
+        message = concept_definition_error.message
+        code = concept_definition_error.structure_class_python_code or ""
+        highlight_lines: set[int] | None = None
+        if syntax_error_data.lineno:
+            highlight_lines = {syntax_error_data.lineno}
+        syntax = Syntax(
+            code=code,
+            lexer="python",
+            line_numbers=True,
+            word_wrap=False,
+            # theme="monokai",  # pick any theme rich knows; omit to use default
+            theme="ansi_dark",  # pick any theme rich knows; omit to use default
+            line_range=None,
+            highlight_lines=highlight_lines,
+        )
+        pretty_range = ""
+        if syntax_error_data.lineno and syntax_error_data.end_lineno:
+            pretty_range = f"lines {syntax_error_data.lineno} to {syntax_error_data.end_lineno}"
+        elif syntax_error_data.lineno:
+            pretty_range = f"line {syntax_error_data.lineno}"
+        if syntax_error_data.offset and syntax_error_data.end_offset:
+            pretty_range += f", column {syntax_error_data.offset} to {syntax_error_data.end_offset}"
+        elif syntax_error_data.offset:
+            pretty_range += f", column {syntax_error_data.offset}"
+        console.print(message)
+        if pretty_range:
+            pretty_print(f"Generated code error at {pretty_range}")
+        console.print(syntax)
