+++ pipelex/core/validation_errors.py
@@ -1,0 +1,7 @@
+from typing_extensions import Protocol
+
+from pipelex.builder.validation_error_data import ConceptDefinitionErrorData
+
+
+class ValidationErrorDetailsProtocol(Protocol):
+    def get_concept_definition_errors(self) -> list[ConceptDefinitionErrorData]: ...
