+++ pipelex/pipe_controllers/sub_pipe_blueprint.py
@@ -1,6 +1,6 @@
 from pydantic import BaseModel, ConfigDict, model_validator
 
-from pipelex.exceptions import PipeDefinitionError
+from pipelex.core.pipe_errors import PipeDefinitionError
 from pipelex.tools.typing.validation_utils import has_more_than_one_among_attributes_from_list
 from pipelex.types import Self
