+++ pipelex/pipe_operators/extract/pipe_extract.py
@@ -13,6 +13,7 @@ from pipelex.cogt.models.model_deck_check import check_extract_choice_with_deck
 from pipelex.config import StaticValidationReaction, get_config
 from pipelex.core.concepts.concept_native import NativeConceptCode
 from pipelex.core.memory.working_memory import WorkingMemory
+from pipelex.core.pipe_errors import PipeDefinitionError
 from pipelex.core.pipes.input_requirements import InputRequirements
 from pipelex.core.pipes.pipe_output import PipeOutput
 from pipelex.core.stuffs.image_content import ImageContent
@@ -22,7 +23,6 @@ from pipelex.core.stuffs.stuff_factory import StuffFactory
 from pipelex.core.stuffs.text_and_images_content import TextAndImagesContent
 from pipelex.core.stuffs.text_content import TextContent
 from pipelex.exceptions import (
-    PipeDefinitionError,
     StaticValidationError,
     StaticValidationErrorType,
 )
