+++ tests/unit/pipelex/core/bundles/test_data_pipe_sorter.py
@@ -3,7 +3,7 @@
 from typing import ClassVar
 
 from pipelex.core.bundles.pipelex_bundle_blueprint import PipeBlueprintUnion
-from pipelex.exceptions import PipeDefinitionError
+from pipelex.core.pipe_errors import PipeDefinitionError
 from pipelex.pipe_controllers.batch.pipe_batch_blueprint import PipeBatchBlueprint
 from pipelex.pipe_controllers.condition.pipe_condition_blueprint import PipeConditionBlueprint
 from pipelex.pipe_controllers.parallel.pipe_parallel_blueprint import PipeParallelBlueprint
