+++ tests/unit/pipelex/pipe_controllers/parallel/test_pipe_parallel_input.py
@@ -3,7 +3,8 @@ from typing import Any
 import pytest
 
 from pipelex import log
-from pipelex.exceptions import PipeDefinitionError, StaticValidationError
+from pipelex.core.pipe_errors import PipeDefinitionError
+from pipelex.exceptions import StaticValidationError
 from pipelex.pipe_controllers.parallel.pipe_parallel_blueprint import PipeParallelBlueprint
 from pipelex.pipe_controllers.parallel.pipe_parallel_factory import PipeParallelFactory
 from tests.unit.pipelex.pipe_controllers.parallel.data import PipeParallelInputTestCases
